+++ NEWS.md
@@ -5,6 +5,7 @@
 New:
 
 * Support for `glmtoolbox::glmgee()` and `glmtoolbox::gnm()` models. Thanks to @luifrancgom for report #1148.
+* Support for `nestedLogit::nestedLogit()` models. Thanks to @strengejacke for report #1675.
 
 Bug fixes: